+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-# 3.0.4
-- cleanup README
+# 3.1.0
+- remove subtraction
+- clean up add/mult -- don't collapse lists to ints
+  - `[1,4,5] + 2` => `[1,4,5,2]`
+  - `[1,4,5] * 2` => `[2,8,10]`
+- remove subtraction -- like division, too many odd error cases
+- clean up error handling -- throw less often, less complex if/else statements.
 
 # 3.0.3
 - update linter and fix analysis